Implement links and modules.
Tickets can have links attached to them (URLs or Attachments). These are modules that can be enabled/disabled.

The idea is if a user submits an "attachment" the attachment is copied over to the attachment folder for an inline-copy (incase the original is deleted, but eventually that might become a setting?)